pyproject.toml
[project]
name = "nj-scraper"
version = "0.1.0"
description = "NJ State Job Scraper"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests==2.31.0",
    "beautifulsoup4==4.12.3",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
